// web/src/app/api/model/markup-request.ts
/**
 * FBC Code Review
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { MarkupGeometry } from './markup-geometry';


export interface MarkupRequest { 
    colour?: string;
    comment?: string;
    /**
     * The finding this markup is about, when it is about one.
     */
    finding_fid?: string;
    geometry: MarkupGeometry;
    kind: MarkupRequestKindEnum;
    page: number;
    sheet?: string;
}
export enum MarkupRequestKindEnum {
    Highlight = 'highlight',
    Box = 'box',
    Arrow = 'arrow',
    Strikeout = 'strikeout',
    Freehand = 'freehand',
    Note = 'note'
};